Insurance Aging Report
This is an easy to read insurance aging report by plan name that will show outstanding claim balance for each plan in the appropriate aging category (by date entered). This also contains total percentages to be utilized in the report designer to show what % of outstanding claim balance is in each aging category.